requirements.txt
absl-py==2.2.2
adk==0.0.5
agno==1.3.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
arxiv==2.2.0
attrs==24.3.0
Authlib==1.6.0
beautifulsoup4==4.13.3
biopython==1.85
blinker==1.9.0
boto3==1.36.18
botocore==1.36.18
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.1
cryptography==45.0.4
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
docstring_parser==0.16
duckduckgo_search==8.0.0
eval_type_backport==0.2.2
faiss-cpu==1.10.0
fastapi==0.115.13
feedparser==6.0.11
filelock==3.17.0
filetype==1.2.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-adk==1.4.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.98.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.25.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
googlesearch-python==1.3.0
graphviz==0.21
greenlet==3.1.1
groq==0.16.0
grpc-google-iam-v1==0.14.2
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.5
jiter==0.10.0
jiwer==3.1.0
jmespath==1.0.1
joblib==1.5.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.66
langchain-google-genai==2.0.9
langchain-mistralai==0.2.6
langchain-ollama==0.2.2
langchain-text-splitters==0.3.5
langsmith==0.3.45
litellm==1.73.1
lxml==5.3.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
mcp==1.9.4
mdurl==0.1.2
mistralai==0.4.2
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.21.1
nltk==3.9.1
numpy==2.2.1
ollama==0.4.7
ollama-ocr==0.1.3
openai==1.91.0
opencv-python==4.11.0.86
opentelemetry-api==1.34.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
phidata==2.7.10
pillow==10.4.0
primp==0.14.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
psycopg2==2.9.10
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.7.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyMuPDF==1.25.1
pyparsing==3.2.1
pypdf==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==13.9.4
rouge_score==0.1.2
rpds-py==0.22.3
rsa==4.9
s3transfer==0.11.2
safetensors==0.5.2
sgmllib3k==1.0.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
sse-starlette==2.3.6
starlette==0.46.2
streamlit==1.41.1
tabulate==0.9.0
tavily-python==0.7.8
tenacity==8.5.0
tiktoken==0.9.0
together==1.3.14
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.1
typer==0.15.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3.1
unzip==1.0.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
watchfiles==1.0.4
websockets==15.0.1
wordcloud==1.9.4
yarl==1.18.3
zipfile36==0.1.3
zipp==3.23.0
zstandard==0.23.0
